requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
numpy==1.23.5
pandas==1.5.3
flask-cors==4.0.0
matplotlib==3.7.0
scikit-learn==1.2.1
fuzzytm>=0.4.0 